# orders-rest-api/orders-api-swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Orders REST API",
        "description": "REST API for managing orders (CRUD operations)",
        "version": "1.0"
    },
    "servers": [{
        "url": "https://apim-placeholder.azure-api.net/orders-api"
    }],
    "paths": {
        "/orders": {
            "get": {
                "summary": "List Orders",
                "description": "List all orders with optional filtering",
                "operationId": "list-orders",
                "parameters": [
                    {
                        "name": "status_filter",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": ["pending", "processing", "shipped", "delivered", "cancelled"]
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "summary": "Create Order",
                "description": "Create a new order",
                "operationId": "create-order",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/orders/{order_id}": {
            "get": {
                "summary": "Get Order",
                "description": "Get order by ID",
                "operationId": "get-order",
                "parameters": [{
                    "name": "order_id",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "summary": "Update Order",
                "description": "Update an existing order",
                "operationId": "update-order",
                "parameters": [{
                    "name": "order_id",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "string"
                    }
                }],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "summary": "Delete Order",
                "description": "Delete an order",
                "operationId": "delete-order",
                "parameters": [{
                    "name": "order_id",
                    "in": "path",
                    "required": true,
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "OrderCreate": {
                "type": "object",
                "required": ["customer_id", "customer_name", "customer_email", "items", "shipping_address"],
                "properties": {
                    "customer_id": {
                        "type": "string"
                    },
                    "customer_name": {
                        "type": "string"
                    },
                    "customer_email": {
                        "type": "string"
                    },
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderItem"
                        }
                    },
                    "shipping_address": {
                        "type": "string"
                    },
                    "notes": {
                        "type": "string"
                    }
                }
            },
            "OrderUpdate": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": ["pending", "processing", "shipped", "delivered", "cancelled"]
                    },
                    "shipping_address": {
                        "type": "string"
                    },
                    "notes": {
                        "type": "string"
                    }
                }
            },
            "OrderItem": {
                "type": "object",
                "required": ["product_id", "product_name", "quantity", "unit_price", "total_price"],
                "properties": {
                    "product_id": {
                        "type": "string"
                    },
                    "product_name": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "unit_price": {
                        "type": "number"
                    },
                    "total_price": {
                        "type": "number"
                    }
                }
            }
        },
        "securitySchemes": {
            "apiKeyHeader": {
                "type": "apiKey",
                "name": "Ocp-Apim-Subscription-Key",
                "in": "header"
            },
            "apiKeyQuery": {
                "type": "apiKey",
                "name": "Ocp-Apim-Subscription-Key",
                "in": "query"
            }
        }
    },
    "security": [{
        "apiKeyHeader": []
    }, {
        "apiKeyQuery": []
    }]
}